feat((export)): reactinterpret export as default
Only one method is exported

BREAKING CHANGE:
chlbri
